reorganized, started working on documentation section
