+++ frontend/src/routes/download/sections/CoordinatesInput.tsx
